requirements.txt
# Core Data Processing and ML
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Synthetic Data Generation
sdv>=0.18.0

# Web Framework (for Phase 7)
Flask>=2.0.0
Flask-WTF>=1.0.0

# Testing and Quality
pytest>=6.2.0
pytest-cov>=2.12.0
hypothesis>=6.0.0

# Configuration Management
PyYAML>=5.4.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.60.0
